docs(README) detailed README
Update the README to give visitors a sense of what the project is,
how it relates to Cycle.js and why it exists.
Touch on relation to Cycle.js, and that it is not a fork that
disregards it's inspiration entirely.
Introduce the goals of the project, and explain the difference and
similarites in the 'Goals' section.
Add an API section to show that it is the same as Cycle.js

Closes #14