feat(dispose): add dispose()
Add dispose() to the object returned by run().
dispose() works by adding .until(disposableStream) to all
streams for sinks and sources. disposableStream, being a stream
from a subject, once dispose() is called a value is pushed
which will terminate all of the streams in *sinks* and *sources.
dispose() also recursively terminates all of the subjects in *sinkProxies*